refactored team api handling
