+++ app/controllers/search_controller.rb
@@ -1,9 +1,7 @@
 require 'httparty'
 require 'json'
-require 'team_api'
 
 class SearchController < ApplicationController
-  include TeamApi
   
   def index
 
@@ -13,12 +11,12 @@ class SearchController < ApplicationController
 
   def results
 
-    @recentHires = get_recent_hires();
     @bodyClass = 'layout-search-results';
     url = 'https://jb-lunr-server.apps.cloud.gov/?q=' + params[:q]
     response = HTTParty.get(url)
     @query = params[:q]
     @searchResults = JSON.parse(response);
+
   end
 end
